compartmentalizing logic; fixing broken recursion (TBD)
