+++ sql_parser/sql_parser/logic/where.py
@@ -1,0 +1,8 @@
+
+from sqlparse.sql import Identifier, IdentifierList, Function, Comparison, Case, Where, Parenthesis, Comment
+from sqlparse.tokens import Keyword, Punctuation, CTE, DML, Comparison as Operator
+from sql_parser import utils as u
+
+
+def is_where(token):
+    return isinstance(token, Where)
